deps: update readdirp to 3.6.0
The signature has changed. file.parentDir was also removed (in
<https://github.com/paulmillr/readdirp/commit/b2db33>), but it was
simply defined as the parent of file.path, and can be replaced as
such.

This also addresses deprecation warnings coming from
micromatch@3.1.10 -> snapdragon -> source-map-resolve. This is in fact
the main motivation for this change.

This might be considered a breaking change, as the minimum Node.js
version of readdirp@3.6.0 is 8.10.0.